+++ frameworks/ans/core/src/ans_manager_stub.cpp
@@ -338,7 +338,7 @@ ErrCode AnsManagerStub::HandlePublishToDevice(MessageParcel &data, MessageParcel
 
 ErrCode AnsManagerStub::HandleCancel(MessageParcel &data, MessageParcel &reply)
 {
-    int notificationId = 0;
+    int32_t notificationId = 0;
     if (!data.ReadInt32(notificationId)) {
         ANS_LOGW("[HandleCancel] fail: read notificationId failed");
         return ERR_ANS_PARCELABLE_FAILED;
@@ -506,14 +506,14 @@ ErrCode AnsManagerStub::HandleGetSlotNumAsBundle(MessageParcel &data, MessagePar
         return ERR_ANS_PARCELABLE_FAILED;
     }
 
-    int num = 0;
+    uint64_t num = 0;
     ErrCode result = GetSlotNumAsBundle(bundleOption, num);
     if (!reply.WriteInt32(result)) {
         ANS_LOGW("[HandleGetSlotNumAsBundle] fail: write result failed, ErrCode=%{public}d", result);
         return ERR_ANS_PARCELABLE_FAILED;
     }
 
-    if (!reply.WriteInt32(num)) {
+    if (!reply.WriteUint64(num)) {
         ANS_LOGW("[HandleGetSlotNumAsBundle] fail: write enabled failed, ErrCode=%{public}d", result);
         return ERR_ANS_PARCELABLE_FAILED;
     }
@@ -549,14 +549,14 @@ ErrCode AnsManagerStub::HandleGetActiveNotifications(MessageParcel &data, Messag
 
 ErrCode AnsManagerStub::HandleGetActiveNotificationNums(MessageParcel &data, MessageParcel &reply)
 {
-    int num = 0;
+    uint64_t num = 0;
     ErrCode result = GetActiveNotificationNums(num);
     if (!reply.WriteInt32(result)) {
         ANS_LOGW("[HandleGetActiveNotificationNums] fail: write result failed, ErrCode=%{public}d", result);
         return ERR_ANS_PARCELABLE_FAILED;
     }
 
-    if (!reply.WriteInt32(num)) {
+    if (!reply.WriteUint64(num)) {
         ANS_LOGW("[HandleGetActiveNotificationNums] fail: write num failed");
         return ERR_ANS_PARCELABLE_FAILED;
     }
@@ -671,7 +671,7 @@ ErrCode AnsManagerStub::HandlePublishAsBundle(MessageParcel &data, MessageParcel
 
 ErrCode AnsManagerStub::HandleSetNotificationBadgeNum(MessageParcel &data, MessageParcel &reply)
 {
-    int num = 0;
+    int32_t num = 0;
     if (!data.ReadInt32(num)) {
         ANS_LOGW("[HandleSetNotificationBadgeNum] fail: read notification failed");
         return ERR_ANS_PARCELABLE_FAILED;
@@ -687,7 +687,7 @@ ErrCode AnsManagerStub::HandleSetNotificationBadgeNum(MessageParcel &data, Messa
 
 ErrCode AnsManagerStub::HandleGetBundleImportance(MessageParcel &data, MessageParcel &reply)
 {
-    int importance = 0;
+    int32_t importance = 0;
     ErrCode result = GetBundleImportance(importance);
     if (!reply.WriteInt32(result)) {
         ANS_LOGW("[HandleGetBundleImportance] fail: write result failed, ErrCode=%{public}d", result);
@@ -780,7 +780,7 @@ ErrCode AnsManagerStub::HandleCancelContinuousTaskNotification(MessageParcel &da
         return ERR_ANS_PARCELABLE_FAILED;
     }
 
-    int notificationId = 0;
+    int32_t notificationId = 0;
     if (!data.ReadInt32(notificationId)) {
         ANS_LOGW("[HandleCancelContinuousTaskNotification] fail: read notificationId failed");
         return ERR_ANS_PARCELABLE_FAILED;
@@ -845,7 +845,7 @@ ErrCode AnsManagerStub::HandleRemoveNotification(MessageParcel &data, MessagePar
         return ERR_ANS_PARCELABLE_FAILED;
     }
 
-    int notificationId = 0;
+    int32_t notificationId = 0;
     if (!data.ReadInt32(notificationId)) {
         ANS_LOGW("[HandleRemoveNotification] fail: read notificationId failed");
         return ERR_ANS_PARCELABLE_FAILED;
@@ -1791,7 +1791,7 @@ ErrCode AnsManagerStub::GetSlotGroups(std::vector<sptr<NotificationSlotGroup>> &
     return ERR_INVALID_OPERATION;
 }
 
-ErrCode AnsManagerStub::GetSlotNumAsBundle(const sptr<NotificationBundleOption> &bundleOption, int &num)
+ErrCode AnsManagerStub::GetSlotNumAsBundle(const sptr<NotificationBundleOption> &bundleOption, uint64_t &num)
 {
     ANS_LOGW("AnsManagerStub::GetSlotNumAsBundle called!");
     return ERR_INVALID_OPERATION;
@@ -1809,7 +1809,7 @@ ErrCode AnsManagerStub::GetActiveNotifications(std::vector<sptr<NotificationRequ
     return ERR_INVALID_OPERATION;
 }
 
-ErrCode AnsManagerStub::GetActiveNotificationNums(int &num)
+ErrCode AnsManagerStub::GetActiveNotificationNums(uint64_t &num)
 {
     ANS_LOGW("AnsManagerStub::GetActiveNotificationNums called!");
     return ERR_INVALID_OPERATION;
